Store answers in local storage (#19)
* More images from new RUV source

* Work on Kjorskra
Probably worst code ever made but it works. Did most of it on my way to
Sauðárkrókur in a car, and the rest of it half asleep. Will clean up
tomorrow

* Moved stuff around a bit on the kjorskra page

* Set currentAddress as soon as we get it so that the UI shows as it's
loading

* Huge overhaul of kjorskra

* Sharing for kjorskra

* Top 3 from every party I could find

* Added build dump folder

* Added docs to the build step

* Lossless compression of candidate images

* Excluding package-lock.json

* updated readme with event link

* Redirect /hack to github repo

* Added thingmenn to the readme

* Push state for kjorskra

* changed phrasing in kjorskra

* added a question option

* Fix scroll issues in Collapsable

There would be an unwanted scroll to the top first time the user
collapses an item.

Now it only scrolls (and updates url) when expanding an item.j

* Make history link smarter

Don't process remote links, other protocols and mouse clicks with
with modifiers.

* Added question reqplies from Vidreisn,Flokkur folksins and Vinstri Graen

* Store answers in local storage

Instead of forcing single session. Add reset button if there's stored
data.